updated feedpage to handle the post creation without having to open up a modal. It does it all on the feed page and its at the very top.
