+++ frontend-runitup/src/styles/CreatePostCard.css
@@ -2,6 +2,7 @@
   margin-bottom: 24px;
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
   border-radius: 8px;
+  transition: all 0.3s ease;
 }
 
 .create-post-card .ant-input {
@@ -19,4 +20,8 @@
 .create-post-actions .ant-btn {
   flex: 1;
   margin: 0 8px;
+}
+
+.create-post-preview {
+  cursor: pointer;
 }
